fix(storage): resolve GCS OTel tracer registry and teardown conflicts in system tests, add get method in cache
